Added a modules to refactor the replaceTemplate function
