+++ index.js
@@ -1,7 +1,7 @@
 const fs = require('fs');
 const http = require('http');
 const url = require('url');
-
+const replaceTemplate = require('./modules/replaceTemplate');
 
 ////////////////////////////////////////////
 ////////// FILES ////////////////////////
@@ -30,19 +30,6 @@ const url = require('url');
 
 ////////////////////////////////////////////
 ////////// SERVER  /////////////////////////
-    const replaceTemplate = (temp, product) => {
-        let output = temp.replace(/{%PRODUCTNAME%}/g, product.productNamde);
-        output = output.replace(/{%IMAGE%}/g, product.image)
-        output = output.replace(/{%PRICE%}/g, product.price)
-        output = output.replace(/{%FROM%}/g, product.from)
-        output = output.replace(/{%NUTRIENTS%}/g, product.nutrients)
-        output = output.replace(/{%QUANTITY%}/g, product.quantity)
-        output = output.replace(/{%DESCRIPTION%}/g, product.description)
-        output = output.replace(/{%ID%}/g, product.id)
-
-        if(!product.organic) output = output.replace(/{%NOT_ORGANIC%}/g, 'not-organic')
-        return output;
-}
 
 const templateOverview = fs.readFileSync('./templates/template-overview.html', 'utf-8');
 const templateCard = fs.readFileSync(`${__dirname}/templates/template-card.html`, 'utf-8');
